% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignSelection.R
\name{optSingleCrit}
\alias{optSingleCrit}
\title{Selection of the best design from the Pareto Front for each criterion}
\usage{
optSingleCrit(ar)
}
\arguments{
\item{ar}{A list as the \code{megaAR} list returned by \code{\link[multiDoE]{runTPLS}}.}
}
\value{
A list whose \eqn{i}-th element corresponds to the solution that optimizes
the \eqn{i}-th criterion. Every solution is a list of two elements:
\itemize{
\item \code{score}: Scores vector.
\item \code{solution}: The design matrix.
}
}
\description{
The \code{optSingleCrit} function selects from the Pareto front
those designs that minimize the criteria when considered individually.
}
