% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignSelection.R
\name{plotPareto}
\alias{plotPareto}
\title{Graphical representation of the Pareto Front}
\usage{
plotPareto(ar, x, y, z = NULL, mode = TRUE)
}
\arguments{
\item{ar}{A list as the \code{megaAR} list returned by \code{\link[multiDoE]{runTPLS}}.}

\item{x}{The criterion on the x axis. It can be one of the following: \code{"I",
"Id", "D", "Ds", "A"} and \code{"As"}.}

\item{y}{The criterion on the y axis. It can be one of the following: \code{"I",
"Id", "D", "Ds", "A"} and \code{"As"}.}

\item{z}{The criterion on the z axis. It can be one of the following: \code{"I",
"Id", "D", "Ds", "A"} and \code{"As"}.}

\item{mode}{When \code{mode=True} the function returns a 3D interactive
chart. When \code{mode=False} it returns a 2D chart in which the \code{z} criteria
values are represented by a color scale.}
}
\value{
The Pareto front chart.
}
\description{
\code{plotPareto} returns a graphical representation (at most 3D)
of the Pareto front.
}
