% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSOpt.R
\name{Score}
\alias{Score}
\title{Criteria values for design matrix}
\usage{
Score(msopt, settings)
}
\arguments{
\item{msopt}{A list as returned by the function \link[multiDoE]{MSOpt}.}

\item{settings}{The design matrix for which criteria scores have to be calculated.}
}
\value{
The data frame of criteria and scores.
}
\description{
The \code{Score} function returns the optimization criteria values for a
given \code{\link[multiDoE]{MSOpt}} list and design matrix.
}
